requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
postgres==4.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3